+++ frontend/app/server-actions/profiles/actions.ts
@@ -4,11 +4,11 @@ import { BattleStadiumApiClient, defaultConfig } from "@/lib/api";
 import { paths } from "@/lib/api/openapi-v1";
 import { FetchOptions } from "openapi-fetch";
 
-export async function getUserProfiles(options?: FetchOptions<paths["/profiles"]["get"]>) {
-  const userProfilesOptions = {
+export async function getProfiles (options?: FetchOptions<paths["/profiles"]["get"]>) {
+  const profilesOptions = {
     ...defaultConfig("getPlayerProfiles"),
     ...options,
   };
 
-  return BattleStadiumApiClient().GET("/profiles", userProfilesOptions);
+  return BattleStadiumApiClient().GET("/profiles", profilesOptions);
 }
